% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_by_time.R
\name{count_by_time}
\alias{count_by_time}
\title{count_by_time function}
\usage{
count_by_time(dataset, date_column, floor_unit)
}
\arguments{
\item{dataset, }{the dataset to be analyzed}

\item{date_column, }{name of the Date column on which to be applied the operation}

\item{floor_unit, }{the date unit to be counted}
}
\value{
a N X 2 dataframe with column "time" and "count", which represents the list of date unit values and the number of their occurrence, respectively.
}
\description{
count_by_time function
}
\details{
Given a Date column, count the occurrence of a date unit specified by floor_unit. This could be useful when finding the pattern between values and specified time frames.
}
\examples{
library(lubridate)
library(dplyr)
library(datateachr)
library(tidyr)
count_by_time(datateachr::vancouver_trees, "date_planted", "year")
count_by_time(datateachr::vancouver_trees, "date_planted", "month")
count_by_time(datateachr::vancouver_trees, "date_planted", "day")

}
